fix(#225): 로그인 모달창 버그 수정
