+++ src/app/(auth)/@modal/(.)welcome/[nickname]/page.tsx
