+++ src/app/layout.tsx
@@ -25,10 +25,8 @@ export const metadata: Metadata = {
 
 export default async function RootLayout({
   children,
-  modal,
 }: {
   children: React.ReactNode;
-  modal: React.ReactNode;
 }) {
   const cookieStore = cookies();
   const user = cookieStore.get('userAccessToken')?.value;
@@ -81,7 +79,7 @@ export default async function RootLayout({
         />
         <main className="relative flex-1">
           {children}
-          {modal}
+          {/* {modal} */}
           <div className="fixed bottom-24 right-8 z-40">
             <ControlOptions />
           </div>
